Implement the road sign module redesign
