+++ app/components/reactive-table.hbs
@@ -18,7 +18,7 @@
       {{yield to="menu"}}
     </menu.general>
   </table.menu>
-  <table.content as |content|>
+  <table.content class={{@tableClass}} as |content|>
     <content.header>
       {{yield (hash
                 Sortable=(
