+++ app/components/related-roadsigns-table.hbs
@@ -8,21 +8,21 @@
     <menu.general>
       <AuToolbar class="au-o-box" as |Group|>
         <Group>
-          <AuHeading @level="2" @skin="3">
+          <AuHeading @level="2" @skin="4">
             {{t "road-marking-concept.attr.related-sign"}}
           </AuHeading>
         </Group>
         <Group>
-          <AuButton {{on "click" @onClickAdd}}>
+          <AuButton @icon="add" {{on "click" @onClickAdd}}>
             {{t "related-road-sign.crud.add"}}
           </AuButton>
         </Group>
       </AuToolbar>
     </menu.general>
   </s.menu>
-  <s.content as |c|>
+  <s.content class="table-centered-content" as |c|>
     <c.header>
-      <th class="data-table__header-title">
+      <th class="data-table__header-title w-px">
         {{t "road-sign-concept.attr.label"}}
       </th>
       <th class="data-table__header-title">
@@ -42,7 +42,12 @@
     </c.header>
     <c.body as |relatedRoadSignConcept|>
       <td>
-        {{relatedRoadSignConcept.label}}
+        <AuLink
+          @route="road-sign-concepts.road-sign-concept"
+          @model={{relatedRoadSignConcept.id}}
+        >
+          {{relatedRoadSignConcept.label}}
+        </AuLink>
       </td>
       <td>
         <img
@@ -63,13 +68,11 @@
       </td>
       <td class="w-px au-u-padding-right-small">
         <AuButton
-          @icon="bin"
           @alert={{true}}
-          @skin="secondary"
-          @hideText={{true}}
+          @skin="naked"
           {{on "click" (fn @onClickRemove relatedRoadSignConcept)}}
         >
-          {{t "related-road-sign.crud.delete"}}
+          {{t "utility.delete"}}
         </AuButton>
       </td>
     </c.body>
