+++ app/components/road-sign-concept/header.hbs
@@ -1,0 +1,80 @@
+<header ...attributes>
+  <AuToolbar @size="large" as |Group|>
+    <Group>
+      <AuHeading @skin="2">
+        {{t "road-sign-concept.details" label=@roadSignConcept.label}}
+      </AuHeading>
+    </Group>
+    <Group>
+      <AuLink
+        @skin="button-secondary"
+        @route="road-sign-concepts.edit"
+        @model={{@roadSignConcept.id}}
+      >
+        {{t "road-sign-concept.crud.edit"}}
+      </AuLink>
+      <AuButton
+        @skin="secondary"
+        @alert={{true}}
+        {{on "click" (fn (mut this.isDeleteConfirmationOpen) true)}}
+      >
+        {{t "road-sign-concept.crud.delete"}}
+      </AuButton>
+
+      <AuModal
+        @modalOpen={{this.isDeleteConfirmationOpen}}
+        @closeModal={{fn (mut this.isDeleteConfirmationOpen) false}}
+      >
+        <:title>
+          {{t "utility.confirmation.title"}}
+        </:title>
+        <:body>
+          <p>
+            {{t "utility.confirmation.body"}}
+          </p>
+        </:body>
+        <:footer>
+          <AuButton
+            @alert={{true}}
+            {{on "click" (perform this.removeRoadSignConcept @roadSignConcept)}}
+          >
+            {{t "road-sign-concept.crud.delete"}}
+          </AuButton>
+          <AuButton
+            @skin="secondary"
+            {{on "click" (fn (mut this.isDeleteConfirmationOpen) false)}}
+          >
+            {{t "utility.cancel"}}
+          </AuButton>
+        </:footer>
+      </AuModal>
+    </Group>
+  </AuToolbar>
+  <AuTabs as |Tab|>
+    <Tab>
+      <AuLink @route="road-sign-concepts.road-sign-concept.instructions">
+        {{t "utility.instructions"}}
+      </AuLink>
+    </Tab>
+    {{#if
+      (and (not @roadSignConcept.isDeleted) (is-sub-sign @roadSignConcept))
+    }}
+      <Tab>
+        <AuLink @route="road-sign-concepts.road-sign-concept.main-signs">
+          {{t "utility.main-signs"}}
+        </AuLink>
+      </Tab>
+    {{else}}
+      <Tab>
+        <AuLink @route="road-sign-concepts.road-sign-concept.sub-signs">
+          {{t "utility.sub-signs"}}
+        </AuLink>
+      </Tab>
+    {{/if}}
+    <Tab>
+      <AuLink @route="road-sign-concepts.road-sign-concept.related">
+        {{t "utility.related-traffic-signs"}}
+      </AuLink>
+    </Tab>
+  </AuTabs>
+</header>
