+++ app/components/sidebar-header.hbs
@@ -1,0 +1,20 @@
+<div class="au-o-box au-c-action-sidebar__header">
+  <AuToolbar class="au-u-margin-bottom-small" as |Group|>
+    <Group>
+      <AuHeading @level="2" @skin="3">
+        {{yield to="title"}}
+      </AuHeading>
+    </Group>
+    <button
+      type="button"
+      class="au-c-close au-c-close--large"
+      {{on "click" @onClose}}
+    >
+      <AuIcon @icon="cross" @size="large" />
+      <span class="au-u-hidden-visually">
+        {{t "utility.close"}}
+      </span>
+    </button>
+  </AuToolbar>
+  {{yield to="content"}}
+</div>
