+++ app/components/validation-status.hbs
@@ -1,0 +1,3 @@
+<AuPill @skin={{if @valid "success" "warning"}}>
+  {{if @valid (t "validation-status.valid") (t "validation-status.draft")}}
+</AuPill>
