+++ app/routes/road-sign-concepts/road-sign-concept/instructions/edit.ts
@@ -1,32 +1,32 @@
-import Store from '@ember-data/store';
+// eslint-disable-next-line ember/use-ember-data-rfc-395-imports -- https://github.com/ember-cli/eslint-plugin-ember/issues/2156
+import type Store from 'ember-data/store';
 import Route from '@ember/routing/route';
-import { inject as service } from '@ember/service';
+import { service } from '@ember/service';
 import type Template from 'mow-registry/models/template';
-import RoadsignConcept from 'mow-registry/routes/road-sign-concepts/road-sign-concept';
+import type RoadsignConcept from 'mow-registry/routes/road-sign-concepts/road-sign-concept';
 import type { ModelFrom } from 'mow-registry/utils/type-utils';
-const PARENT_ROUTE = 'road-sign-concepts.road-sign-concept';
 
 type Params = {
-  instruction_id: string;
+  instructionId: string;
 };
 
-export default class RoadSignConceptsRoadSignConceptInstructionRoute extends Route {
+export default class RoadSignConceptsRoadSignConceptInstructionsEditRoute extends Route {
   @service declare store: Store;
 
   async model(params: Params) {
-    const parentModel = this.modelFor(
-      PARENT_ROUTE,
+    const { roadSignConcept } = this.modelFor(
+      'road-sign-concepts.road-sign-concept',
     ) as ModelFrom<RoadsignConcept>;
-    const concept = parentModel.roadSignConcept;
-    if (params.instruction_id === 'new') {
+    const concept = roadSignConcept;
+    if (params.instructionId === 'new') {
       return {
+        roadSignConcept,
         concept,
-        from: PARENT_ROUTE,
       };
     } else {
       const template = await this.store.findRecord<Template>(
         'template',
-        params.instruction_id,
+        params.instructionId,
         {
           // @ts-expect-error we're running into strange type errors with the query argument. Not sure how to fix this properly.
           // TODO: fix the query types
@@ -35,8 +35,8 @@ export default class RoadSignConceptsRoadSignConceptInstructionRoute extends Rou
       );
       return {
         template,
+        roadSignConcept,
         concept,
-        from: PARENT_ROUTE,
       };
     }
   }
