+++ app/styles/app.scss
@@ -41,6 +41,8 @@
 
 // COMPONENTS
 @import "@appuniversum/ember-appuniversum/styles/a-components";
+@import "components/sidebar-container";
+@import "components/sidebar-overlay-container";
 @import "breadcrumbs";
 @import "input-file";
 @import "thumbnail";
