--- app/templates/road-sign-concepts/road-sign-concept/instruction.hbs
@@ -1,6 +1,0 @@
-{{page-title (t "utility.add-instruction")}}
-<AddInstruction
-  @from={{this.model.from}}
-  @editedTemplate={{this.model.template}}
-  @concept={{this.model.concept}}
-/>
