+++ app/templates/road-sign-concepts/road-sign-concept/instructions.hbs
@@ -1,0 +1,11 @@
+{{page-title (t "utility.instructions")}}
+<BreadcrumbsItem as |linkClass|>
+  <LinkTo
+    @route="road-sign-concepts.road-sign-concept.instructions"
+    class={{linkClass}}
+  >
+    {{t "utility.instructions"}}
+  </LinkTo>
+</BreadcrumbsItem>
+
+{{outlet}}
