+++ app/templates/road-sign-concepts/road-sign-concept/instructions/index.hbs
@@ -1,0 +1,60 @@
+<RoadSignConcept::Header @roadSignConcept={{@model.roadSignConcept}} />
+<AuDataTable
+  @content={{@model.templates}}
+  @fields="Description"
+  @noDataMessage={{t "road-sign-concept.crud.no-instruction"}}
+  as |s|
+>
+  <s.menu as |menu|>
+    <menu.general>
+      <AuToolbar class="au-o-box" as |Group|>
+        <Group>
+          <AuHeading @level="2" @skin="3" class="au-u-hidden-visually">
+            {{t "utility.instructions"}}
+          </AuHeading>
+        </Group>
+        <Group>
+          <AuLink
+            @route="road-sign-concepts.road-sign-concept.instructions.edit"
+            @model="new"
+            @skin="button"
+            @icon="add"
+          >
+            {{t "utility.add-instruction"}}
+          </AuLink>
+        </Group>
+      </AuToolbar>
+    </menu.general>
+  </s.menu>
+  <s.content as |c|>
+    <c.header>
+      <th class="data-table__header-title">
+        {{t "utility.description"}}
+      </th>
+      <th>
+        <span class="au-u-hidden-visually">
+          {{t "main-sign.crud.delete"}}
+        </span>
+      </th>
+    </c.header>
+    <c.body as |template|>
+      <td>
+        <AuLink
+          @route="road-sign-concepts.road-sign-concept.instructions.edit"
+          @model={{template.id}}
+        >
+          {{template.value}}
+        </AuLink>
+      </td>
+      <td class="w-px au-u-padding-right-small">
+        <AuButton
+          @alert={{true}}
+          @skin="naked"
+          {{on "click" (fn this.removeTemplate.perform template)}}
+        >
+          {{t "utility.delete"}}
+        </AuButton>
+      </td>
+    </c.body>
+  </s.content>
+</AuDataTable>
